@@ -8,7 +8,7 @@ import { StringComponent } from "../string/string";
 import { SortingPage } from "../sorting-page/sorting-page";
 import { StackPage } from "../stack-page/stack-page";
 
-import styles from "./app.module.css"
+import styles from "./app.module.css";
 
 function App() {
   return (
