@@ -37,7 +37,7 @@ export const StringComponent: React.FC = () => {
       setArr(res);
       setStartIndex(start);
       value.length % 2 === 0 ? setEndIndex(end - 1) : setEndIndex(end);
-      await new Promise((resolve) => setTimeout(resolve, 1500));
+      await new Promise((resolve) => setTimeout(resolve, 500));
       value.length % 2 === 0
         ? ([res[start], res[end - 1]] = [res[end - 1], res[start]])
         : ([res[start], res[end]] = [res[end], res[start]]);
@@ -61,6 +61,7 @@ export const StringComponent: React.FC = () => {
           }}
           style={{minWidth: '178px'}}
           disabled={value === ""}
+          id="reverse"
         />
       </div>
       <div className={styles.letters}>
